fix: recursive module
